Eliminate GtkBoxes used as spacers
In preferences_section, a box was being used to indent the items
underneath each section heading. Replace with a `margin_start`
value set on the content box, instead.

In properties_page, an expanding box was being used to separate the
filename from the other file properties — switch to `pack_end()`
when placing the other data labels, so the spacing is automatic.